update: check and controllers + methods + DB ( is 87 string working  ?).

+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Data
 public class Review {
-    long reviewId;
+    int reviewId;
     String content;
     boolean isPositive;
     int userId;
